requirements.txt
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
python-dotenv
pymupdf
langchain-core


